+++ packages/core/src/browser/api.ts
@@ -1,0 +1,8 @@
+/********************************************************************************
+ * Copyright (c) 2026 CrossBreeze.
+ ********************************************************************************/
+import { EditorOpenerOptions } from '@theia/editor/lib/browser';
+
+export interface CompositeEditorOpenerOptions extends EditorOpenerOptions {
+   perspective?: 'code' | 'primary';
+}